Feat: User Story 1 / and Adds button to discout show page

+++ db/migrate/20230303175102_create_bulk_discount.rb
@@ -1,7 +1,7 @@
 class CreateBulkDiscount < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
-      t.float :percentage_discount
+      t.decimal :percentage_discount
       t.integer :quantity_threshhold
       t.references :merchant, foreign_key: true
       t.timestamps
